Added a few more error codes and messages.
Now all errors coming from the library should be ours. Also, if an error occurs during file cleanup, that is sent to the delagate but not saved...so the error we return from a failed zip/unzip is the failure error and not some error that ocurred while trying to cleanup any folders/files we have created.