Make sure [ZipFile close] is only called once.
close can throw exceptions, which means that the dealloc for ZipWithProgress can throw exceptions.  Because of ARC, dealloc can be called at unpredictable times.  If close throws an exception, the ZipFile is in a weird state, and a second call to close can crash.

I do not catch the exception, because I want the leave that responsibility to the client.

things are in a weird state.